Simple generation of software keys using openssl, only available if the 'softkeys' feature is used.
