+++ Cargo.toml
@@ -15,10 +15,10 @@ failure         = "^0.1"
 hex             = "^0.3"
 log             = "^0.4"
 openssl         = { version = "^0.10", optional = true }
-ring            = { version  = "^0.13", features = ["rsa_signing"] }
+ring            = "^0.13"
 untrusted       = "^0.6"
 xml-rs          = "^0.8"
 
 [features]
 default = []
-signing = [ "openssl"]
+softkeys = [ "openssl" ]
